Projeto 1/CMovW1V1/Project_files/CompanyServer/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
pycryptodome==3.7.0
pytz==2018.5
virtualenv==16.0.0
